+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-02T10:04:09.782Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-02T12:26:15.830Z")
     r"""The date and time the resource was last updated"""